+++ .github/workflows/test.yaml
@@ -2,6 +2,9 @@ name: CI
 
 on: [pull_request]
 
+env:
+  CI: true
+
 jobs:
   build:
     runs-on: ubuntu-latest
