+++ e2e/bugs/debounce-open-e2e.spec.ts
@@ -1,34 +1,36 @@
 import {DemoPage} from '../app.po';
-import {browser} from 'protractor';
+import {expect, Page, test} from '@playwright/test';
 
-describe('debounce open delay', () => {
-  let page: DemoPage;
+test.describe('debounce open delay', () => {
+  let po: DemoPage;
+  let page: Page;
 
-  beforeEach(async () => {
-    page = new DemoPage();
-    await page.navigateTo();
+  test.beforeAll(async ({browser}) => {
+    page = await browser.newPage();
   });
 
-  it('should open a picker when there is an openDelay', async () => {
-    await page.setInputValue(page.onOpenDelayInput, '1000');
-    await page.dayPickerInput.click();
-    await browser.waitForAngularEnabled(false);
-    await browser.sleep(500);
-    expect(await page.datePickerPopup.isPresent()).toBe(false);
-    await browser.sleep(600);
-    await browser.waitForAngularEnabled(true);
-    expect(await page.datePickerPopup.isDisplayed()).toBe(true);
+  test.beforeEach(async () => {
+    po = new DemoPage(page);
+    await po.navigateTo();
   });
 
-  it('should cancel open of a picker when there is an openDelay and focus out', async () => {
-    await page.setInputValue(page.onOpenDelayInput, '1000');
-    await page.dayPickerInput.click();
-    await browser.waitForAngularEnabled(false);
-    await browser.sleep(500);
-    expect(await page.datePickerPopup.isPresent()).toBe(false);
-    await page.clickOnBody();
-    await browser.sleep(600);
-    await browser.waitForAngularEnabled(true);
-    expect(await page.datePickerPopup.isPresent()).toBe(false);
+  test('should open a picker when there is an openDelay', async () => {
+    await po.setText(po.onOpenDelayInput(), '1000');
+    await po.dayPickerInput().click();
+    
+    await po.sleep(500);
+    expect(po.datePickerPopup()).toBeHidden();
+    await po.sleep(600);
+    expect(po.datePickerPopup()).toBeVisible();
+  });
+
+  test('should cancel open of a picker when there is an openDelay and focus out', async () => {
+    await po.setText(po.onOpenDelayInput(), '1000');
+    await po.dayPickerInput().click();
+    await po.sleep(500);
+    expect(po.datePickerPopup()).toBeHidden();
+    await po.clickOnBody();
+    await po.sleep(600);
+    expect(po.datePickerPopup()).toBeHidden();
   });
 });
