+++ e2e/bugs/format-change-e2e.spec.ts
@@ -1,58 +1,64 @@
 import {DemoPage} from '../app.po';
+import {expect, Page, test} from '@playwright/test';
 
-describe('format changed validation', () => {
-  let page: DemoPage;
+test.describe('format changed validation', () => {
+  let po: DemoPage;
+  let page: Page;
 
-  beforeEach(async () => {
-    page = new DemoPage();
-    await page.navigateTo();
+  test.beforeAll(async ({browser}) => {
+    page = await browser.newPage();
   });
 
-  it('should hide validation on date time picker', async () => {
-    await page.daytimePickerMenu.click();
-    await page.setInputValue(page.dateFormatInput, 'DD-MM-YYYY');
-    await page.setInputValue(page.daytimePickerInput, '10-04-2017 09:08:07');
-    await page.clickOnBody();
-    expect(await page.formatValidationMsg.isPresent()).toBeTruthy();
-    await page.setInputValue(page.dateFormatInput, 'DD-MM-YYYY HH:mm:ss');
-    await page.setInputValue(page.daytimePickerInput, '10-04-2017 09:08:07');
-    await page.clickOnBody();
-    expect(await page.formatValidationMsg.isPresent()).toBeFalsy();
+  test.beforeEach(async () => {
+    po = new DemoPage(page);
+    await po.navigateTo();
   });
 
-  it('should hide validation on day picker', async () => {
-    await page.dayPickerMenu.click();
-    await page.setInputValue(page.dateFormatInput, 'DD-MM');
-    await page.setInputValue(page.dayPickerInput, '10-04-2017');
-    await page.clickOnBody();
-    expect(await page.formatValidationMsg.isPresent()).toBeTruthy();
-    await page.setInputValue(page.dateFormatInput, 'DD-MM-YYYY');
-    await page.setInputValue(page.dayPickerInput, '10-04-2017');
-    await page.clickOnBody();
-    expect(await page.formatValidationMsg.isPresent()).toBeFalsy();
+  test('should hide validation on date time picker', async () => {
+    await po.daytimePickerMenu().click();
+    await po.setText(po.dateFormatInput(), 'DD-MM-YYYY');
+    await po.setText(po.daytimePickerInput(), '10-04-2017 09:08:07');
+    await po.clickOnBody();
+    expect(po.formatValidationMsg()).toBeVisible();
+    await po.setText(po.dateFormatInput(), 'DD-MM-YYYY HH:mm:ss');
+    await po.setText(po.daytimePickerInput(), '10-04-2017 09:08:07');
+    await po.clickOnBody();
+    expect(po.formatValidationMsg()).toBeHidden();
   });
 
-  it('should work on time picker picker', async () => {
-    await page.timePickerMenu.click();
-    await page.setInputValue(page.dateFormatInput, 'HH');
-    await page.setInputValue(page.timePickerInput, '10:04');
-    await page.clickOnBody();
-    expect(await page.formatValidationMsg.isPresent()).toBeTruthy();
-    await page.setInputValue(page.dateFormatInput, 'HH:mm');
-    await page.timePickerInput.click();
-    await page.clickOnBody();
-    expect(await page.formatValidationMsg.isPresent()).toBeFalsy();
+  test('should hide validation on day picker', async () => {
+    await po.dayPickerMenu().click();
+    await po.setText(po.dateFormatInput(), 'DD-MM');
+    await po.setText(po.dayPickerInput(), '10-04-2017');
+    await po.clickOnBody();
+    expect(po.formatValidationMsg()).toBeVisible();
+    await po.setText(po.dateFormatInput(), 'DD-MM-YYYY');
+    await po.setText(po.dayPickerInput(), '10-04-2017');
+    await po.clickOnBody();
+    expect(po.formatValidationMsg()).toBeHidden();
   });
 
-  it('should hide validation on month picker', async () => {
-    await page.monthDirectiveMenu.click();
-    await page.setInputValue(page.dateFormatInput, 'MM');
-    await page.setInputValue(page.monthDirectiveInput, 'Jan');
-    await page.clickOnBody();
-    expect(await page.formatValidationMsg.isPresent()).toBeTruthy();
-    await page.setInputValue(page.dateFormatInput, 'MMM');
-    await page.setInputValue(page.monthDirectiveInput, 'Jan');
-    await page.clickOnBody();
-    expect(await page.formatValidationMsg.isPresent()).toBeFalsy();
+  test('should work on time picker picker', async () => {
+    await po.timePickerMenu().click();
+    await po.setText(po.dateFormatInput(), 'HH');
+    await po.setText(po.timePickerInput(), '10:04');
+    await po.clickOnBody();
+    expect(po.formatValidationMsg()).toBeVisible();
+    await po.setText(po.dateFormatInput(), 'HH:mm');
+    await po.timePickerInput().click();
+    await po.clickOnBody();
+    expect(po.formatValidationMsg()).toBeHidden();
+  });
+
+  test('should hide validation on month picker', async () => {
+    await po.monthDirectiveMenu().click();
+    await po.setText(po.dateFormatInput(), 'MM');
+    await po.setText(po.monthDirectiveInput(), 'Jan');
+    await po.clickOnBody();
+    expect(po.formatValidationMsg()).toBeVisible();
+    await po.setText(po.dateFormatInput(), 'MMM');
+    await po.setText(po.monthDirectiveInput(), 'Jan');
+    await po.clickOnBody();
+    expect(po.formatValidationMsg()).toBeHidden();
   });
 });
