+++ e2e/fetures/locale-e2e.spec.ts
@@ -1,54 +1,60 @@
 import {DemoPage} from '../app.po';
+import {expect, Page, test} from '@playwright/test';
 
-describe('Locales', () => {
-  let page: DemoPage;
+test.describe('Locales', () => {
+  let po: DemoPage;
+  let page: Page;
 
-  beforeEach(async () => {
-    page = new DemoPage();
-    await page.navigateTo();
+  test.beforeAll(async ({browser}) => {
+    page = await browser.newPage();
   });
 
-  it('should check day time picker locale', async () => {
-    await page.daytimePickerMenu.click();
-    await page.hebrewLocale.click();
-    await page.daytimePickerInput.click();
-    expect(await page.weekDayNames.getText()).toEqual(['א׳ב׳ג׳ד׳ה׳ו׳ש׳']);
+  test.beforeEach(async () => {
+    po = new DemoPage(page);
+    await po.navigateTo();
+  });
 
-    await page.daytimeInlineMenu.click();
-    await page.hebrewLocale.click();
-    expect(await page.weekDayInline.getText()).toEqual(['א׳ב׳ג׳ד׳ה׳ו׳ש׳']);
+  test('should check day time picker locale', async () => {
+    await po.daytimePickerMenu().click();
+    await po.localeSelect().selectOption('he');
+    await po.daytimePickerInput().click();
+    expect(await po.weekDayNames().allTextContents()).toEqual(['א׳ב׳ג׳ד׳ה׳ו׳ש׳']);
 
-    await page.daytimeDirectiveMenu.click();
-    await page.hebrewLocale.click();
-    await page.daytimeDirectiveInput.click();
-    expect(await page.weekDayNames.getText()).toEqual(['א׳ב׳ג׳ד׳ה׳ו׳ש׳']);
+    await po.daytimeInlineMenu().click();
+    await po.localeSelect().selectOption('he');
+    expect(await po.weekDayInline().allTextContents()).toEqual(['א׳ב׳ג׳ד׳ה׳ו׳ש׳']);
 
-    await page.dayPickerMenu.click();
-    await page.hebrewLocale.click();
-    await page.dayPickerInput.click();
-    expect(await page.weekDayNames.getText()).toEqual(['א׳ב׳ג׳ד׳ה׳ו׳ש׳']);
+    await po.daytimeDirectiveMenu().click();
+    await po.localeSelect().selectOption('he');
+    await po.daytimeDirectiveInput().click();
+    expect(await po.weekDayNames().allTextContents()).toEqual(['א׳ב׳ג׳ד׳ה׳ו׳ש׳']);
 
-    await page.dayInlineMenu.click();
-    await page.hebrewLocale.click();
-    expect(await page.weekDayInline.getText()).toEqual(['א׳ב׳ג׳ד׳ה׳ו׳ש׳']);
+    await po.dayPickerMenu().click();
+    await po.localeSelect().selectOption('he');
+    await po.dayPickerInput().click();
+    expect(await po.weekDayNames().allTextContents()).toEqual(['א׳ב׳ג׳ד׳ה׳ו׳ש׳']);
 
-    await page.dayDirectiveMenu.click();
-    await page.hebrewLocale.click();
-    await page.dayDirectiveInput.click();
-    expect(await page.weekDayNames.getText()).toEqual(['א׳ב׳ג׳ד׳ה׳ו׳ש׳']);
+    await po.dayInlineMenu().click();
+    await po.localeSelect().selectOption('he');
+    expect(await po.weekDayInline().allTextContents()).toEqual(['א׳ב׳ג׳ד׳ה׳ו׳ש׳']);
 
-    await page.monthPickerMenu.click();
-    await page.hebrewLocale.click();
-    await page.monthPickerInput.click();
-    expect(await page.calendarFirstMonthOfYear.getText()).toEqual('ינו');
+    await po.dayDirectiveMenu().click();
+    await po.localeSelect().selectOption('he');
+    await po.dayDirectiveInput().click();
+    expect(await po.weekDayNames().allTextContents()).toEqual(['א׳ב׳ג׳ד׳ה׳ו׳ש׳']);
 
-    await page.monthInlineMenu.click();
-    await page.hebrewLocale.click();
-    expect(await page.calendarFirstMonthOfYearInline.getText()).toEqual('ינו');
+    await po.monthPickerMenu().click();
+    await po.localeSelect().selectOption('he');
+    await po.monthPickerInput().click();
+    expect(await po.calendarFirstMonthOfYear().first().textContent()).toEqual('ינו');
 
-    await page.monthDirectiveMenu.click();
-    await page.hebrewLocale.click();
-    await page.monthDirectiveInput.click();
-    expect(await page.calendarFirstMonthOfYear.getText()).toEqual('ינו');
+    await po.monthInlineMenu().click();
+    await po.localeSelect().selectOption('he');
+    expect(await po.calendarFirstMonthOfYearInline().first().textContent()).toEqual('ינו');
+
+    await po.monthDirectiveMenu().click();
+    await po.localeSelect().selectOption('he');
+    await po.monthDirectiveInput().click();
+    expect(await po.calendarFirstMonthOfYear().first().textContent()).toEqual('ינו');
   });
 });
