+++ e2e/timepicker-e2e.spec.ts
@@ -1,131 +1,136 @@
 import {DemoPage} from './app.po';
+import {expect, Page, test} from '@playwright/test';
 
-describe('dpDayPicker timePicker', () => {
-  let page: DemoPage;
+test.describe('dpDayPicker timePicker', () => {
+  let po: DemoPage;
+  let page: Page;
 
-  beforeEach(async () => {
-    page = new DemoPage();
-    await page.navigateTo();
-
-    await page.timePickerMenu.click();
-    await page.clickOnBody();
+  test.beforeAll(async ({browser}) => {
+    page = await browser.newPage();
   });
 
-  it('should check if min time validation is working', async () => {
-    expect(await page.minDateValidationMsg.isPresent()).toBe(false);
-    await page.setInputValue(page.minTimeValidationPickerInput, '10:00:00');
-    await page.setInputValue(page.timePickerInput, '09:00:00');
-    await page.clickOnBody();
-    expect(await page.minDateValidationMsg.getText()).toEqual('minDate invalid');
-    await page.setInputValue(page.minTimeValidationPickerInput, '08:07:06');
-    await page.clickOnBody();
-    expect(await page.minDateValidationMsg.isPresent()).toBe(false);
+  test.beforeEach(async () => {
+    po = new DemoPage(page);
+    await po.navigateTo();
+    await po.timePickerMenu().click();
+    await po.clickOnBody();
   });
 
-  it('should check if max time validation is working', async () => {
-    page.maxTimeValidationPickerInput.click();
-    page.maxTimeValidationPickerInput.clear();
-    expect(await page.maxDateValidationMsg.isPresent()).toBe(false);
-    page.maxTimeValidationPickerInput.clear();
-    page.maxTimeValidationPickerInput.sendKeys('08:07:06');
-    page.timePickerInput.click();
-    page.timePickerInput.clear();
-    page.timePickerInput.sendKeys('09:00:00');
-    page.clickOnBody();
-    expect(await page.maxDateValidationMsg.getText()).toEqual('maxDate invalid');
-    page.maxTimeValidationPickerInput.clear();
-    page.maxTimeValidationPickerInput.sendKeys('10:11:12');
-    expect(await page.maxDateValidationMsg.isPresent()).toBe(false);
+  test('should check if min time validation is working', async () => {
+    await expect(po.minDateValidationMsg()).toBeHidden();
+    await po.setText(po.minTimeValidationPickerInput(), '10:00:00');
+    await po.setText(po.timePickerInput(), '09:00:00');
+    await po.clickOnBody();
+    expect(await po.minDateValidationMsg().textContent()).toEqual('minDate invalid');
+    await po.setText(po.minTimeValidationPickerInput(), '08:07:06');
+    await po.clickOnBody();
+    await expect(po.minDateValidationMsg()).toBeHidden();
   });
 
-  it('should check that the min selectable time option is working', async () => {
-    page.minTimeInput.click();
-    page.minTimeInput.clear();
-    page.minTimeInput.sendKeys('08:07:06');
-    page.timePickerInput.click();
-    page.timePickerInput.clear();
-    page.timePickerInput.sendKeys('09:00:00');
-    expect(await page.hourDownBtn.getAttribute('disabled')).toEqual('true');
-    expect(await page.minuteDownBtn.getAttribute('disabled')).toBe(null);
-    expect(await page.meridiemUpBtn.getAttribute('disabled')).toBe(null);
-    expect(await page.meridiemDownBtn.getAttribute('disabled')).toBe(null);
+  test('should check if max time validation is working', async () => {
+    await po.maxTimeValidationPickerInput().click();
+    await po.maxTimeValidationPickerInput().clear();
+    await expect(po.maxDateValidationMsg()).toBeHidden();
+    await po.maxTimeValidationPickerInput().clear();
+    await po.setText(po.maxTimeValidationPickerInput(), '08:07:06');
+    await po.timePickerInput().click();
+    await po.timePickerInput().clear();
+    await po.setText(po.timePickerInput(), '09:00:00');
+    await po.clickOnBody();
+    expect(await po.maxDateValidationMsg().textContent()).toEqual('maxDate invalid');
+    await po.maxTimeValidationPickerInput().clear();
+    await po.setText(po.maxTimeValidationPickerInput(), '10:11:12');
+    await expect(po.maxDateValidationMsg()).toBeHidden();
   });
 
-  it('should check that the max selectable time option is working', async () => {
-    page.maxTimeInput.click();
-    page.maxTimeInput.clear();
-    page.maxTimeInput.sendKeys('09:07:06');
-    page.timePickerInput.click();
-    page.timePickerInput.clear();
-    page.timePickerInput.sendKeys('09:00:00');
-    expect(await page.hourUpBtn.getAttribute('disabled')).toEqual('true');
-    expect(await page.minuteUpBtn.getAttribute('disabled')).toBe(null);
-    expect(await page.meridiemUpBtn.getAttribute('disabled')).toEqual('true');
-    expect(await page.meridiemDownBtn.getAttribute('disabled')).toEqual('true');
+  test('should check that the min selectable time option is working', async () => {
+    await po.minTimeInput().click();
+    await po.minTimeInput().clear();
+    await po.setText(po.minTimeInput(), '08:07:06');
+    await po.timePickerInput().click();
+    await po.timePickerInput().clear();
+    await po.setText(po.timePickerInput(), '09:00:00');
+    await expect(await po.hourDownBtn()).toBeDisabled();
+    await expect(await po.minuteDownBtn()).not.toBeDisabled();
+    await expect(await po.meridiemUpBtn()).not.toBeDisabled();
+    await expect(await po.meridiemDownBtn()).not.toBeDisabled();
   });
 
-  it('should check that the 12 hour time format options work', async () => {
-    page.showSeconds.click();
-    page.hours12FormatInput.clear();
-    page.hours12FormatInput.sendKeys('h');
-    page.minutesFormatInput.clear();
-    page.minutesFormatInput.sendKeys('m');
-    page.secondsFormatInput.clear();
-    page.secondsFormatInput.sendKeys('s');
-    page.meridiemFormatInput.clear();
-    page.meridiemFormatInput.sendKeys('a');
-    page.timeSeparatorInput.clear();
-    page.timeSeparatorInput.sendKeys('-');
-    page.timePickerInput.click();
-    page.timePickerInput.clear();
-    page.timePickerInput.sendKeys('09:08:07');
-    expect(await page.hourDisplay.getText()).toEqual('9');
-    expect(await page.minuteDisplay.getText()).toEqual('8');
-    expect(await page.secondDisplay.getText()).toEqual('7');
-    expect(await page.meridiemDisplay.getText()).toEqual('am');
-    expect(await page.timeSeparatorDisplay.getText()).toEqual('-');
+  test('should check that the max selectable time option is working', async () => {
+    await po.maxTimeInput().click();
+    await po.maxTimeInput().clear();
+    await po.setText(po.maxTimeInput(), '09:07:06');
+    await po.timePickerInput().click();
+    await po.timePickerInput().clear();
+    await po.setText(po.timePickerInput(), '09:00:00');
+    await expect(po.hourUpBtn()).toBeDisabled();
+    await expect(po.minuteUpBtn()).not.toBeDisabled();
+    await expect(po.meridiemUpBtn()).toBeDisabled();
+    await expect(po.meridiemDownBtn()).toBeDisabled();
   });
 
-  it('should check that the 24 hour time format options work', async () => {
-    page.showTwentyFourHours.click();
-    page.showSeconds.click();
-    page.hours24FormatInput.clear();
-    page.hours24FormatInput.sendKeys('H');
-    page.minutesFormatInput.clear();
-    page.minutesFormatInput.sendKeys('m');
-    page.secondsFormatInput.clear();
-    page.secondsFormatInput.sendKeys('s');
-    page.timeSeparatorInput.clear();
-    page.timeSeparatorInput.sendKeys('-');
-    page.timePickerInput.click();
-    page.timePickerInput.clear();
-    page.timePickerInput.sendKeys('09:08:07');
-    expect(await page.hourDisplay.getText()).toEqual('9');
-    expect(await page.minuteDisplay.getText()).toEqual('8');
-    expect(await page.secondDisplay.getText()).toEqual('7');
-    expect(await page.timeSeparatorDisplay.getText()).toEqual('-');
+  test('should check that the 12 hour time format options work', async () => {
+    await po.showSeconds().click();
+    await po.hours12FormatInput().clear();
+    await po.setText(po.hours12FormatInput(), 'h');
+    await po.minutesFormatInput().clear();
+    await po.setText(po.minutesFormatInput(), 'm');
+    await po.secondsFormatInput().clear();
+    await po.setText(po.secondsFormatInput(), 's');
+    await po.meridiemFormatInput().clear();
+    await po.setText(po.meridiemFormatInput(), 'a');
+    await po.timeSeparatorInput().clear();
+    await po.setText(po.timeSeparatorInput(), '-');
+    await po.timePickerInput().click();
+    await po.timePickerInput().clear();
+    await po.setText(po.timePickerInput(), '09:08:07');
+    expect(await po.hourDisplay().textContent()).toEqual('9');
+    expect(await po.minuteDisplay().textContent()).toEqual('8');
+    expect(await po.secondDisplay().textContent()).toEqual('7');
+    expect(await po.meridiemDisplay().textContent()).toEqual('am');
+    expect(await po.timeSeparatorDisplay().textContent()).toEqual('-');
   });
 
-  it('should check that the interval options work', async () => {
-    page.showSeconds.click();
-    page.minutesIntervalInput.clear();
-    page.minutesIntervalInput.sendKeys('6');
-    page.secondsIntervalInput.clear();
-    page.secondsIntervalInput.sendKeys('7');
-    page.timePickerInput.click();
-    page.timePickerInput.clear();
-    page.timePickerInput.sendKeys('09:08:07');
-    page.minuteUpBtn.click();
-    expect(await page.minuteDisplay.getText()).toEqual('14');
-    expect(await page.getInputVal(page.timePickerInput)).toEqual('09:14:07');
-    page.minuteDownBtn.click();
-    expect(await page.minuteDisplay.getText()).toEqual('08');
-    expect(await page.getInputVal(page.timePickerInput)).toEqual('09:08:07');
-    page.secondUpBtn.click();
-    expect(await page.secondDisplay.getText()).toEqual('14');
-    expect(await page.getInputVal(page.timePickerInput)).toEqual('09:08:14');
-    page.secondDownBtn.click();
-    expect(await page.secondDisplay.getText()).toEqual('07');
-    expect(await page.getInputVal(page.timePickerInput)).toEqual('09:08:07');
+  test('should check that the 24 hour time format options work', async () => {
+    await po.showTwentyFourHours().click();
+    await po.showSeconds().click();
+    await po.hours24FormatInput().clear();
+    await po.setText(po.hours24FormatInput(), 'H');
+    await po.minutesFormatInput().clear();
+    await po.setText(po.minutesFormatInput(), 'm');
+    await po.secondsFormatInput().clear();
+    await po.setText(po.secondsFormatInput(), 's');
+    await po.timeSeparatorInput().clear();
+    await po.setText(po.timeSeparatorInput(), '-');
+    await po.timePickerInput().click();
+    await po.timePickerInput().clear();
+    await po.setText(po.timePickerInput(), '09:08:07');
+    expect(await po.hourDisplay().textContent()).toEqual('9');
+    expect(await po.minuteDisplay().textContent()).toEqual('8');
+    expect(await po.secondDisplay().textContent()).toEqual('7');
+    expect(await po.timeSeparatorDisplay().textContent()).toEqual('-');
+  });
+
+  test('should check that the interval options work', async () => {
+    await po.showSeconds().click();
+    await po.minutesIntervalInput().clear();
+    await po.setText(po.minutesIntervalInput(), '6');
+    await po.secondsIntervalInput().clear();
+    await po.setText(po.secondsIntervalInput(), '7');
+    await po.timePickerInput().click();
+    await po.timePickerInput().clear();
+    await po.setText(po.timePickerInput(), '09:08:07');
+    await po.minuteUpBtn().click();
+    expect(await po.minuteDisplay().textContent()).toEqual('14');
+    await expect(po.timePickerInput()).toHaveValue('09:14:07');
+    await po.minuteDownBtn().click();
+    expect(await po.minuteDisplay().textContent()).toEqual('08');
+    await expect(po.timePickerInput()).toHaveValue('09:08:07');
+    await po.secondUpBtn().click();
+    expect(await po.secondDisplay().textContent()).toEqual('14');
+    await expect(po.timePickerInput()).toHaveValue('09:08:14');
+    await po.secondDownBtn().click();
+    expect(await po.secondDisplay().textContent()).toEqual('07');
+    await expect(po.timePickerInput()).toHaveValue('09:08:07');
   });
 });
