--- e2e/tsconfig.e2e.json
@@ -1,19 +1,0 @@
-{
-  "compilerOptions": {
-    "sourceMap": true,
-    "declaration": false,
-    "moduleResolution": "node",
-    "experimentalDecorators": true,
-    "lib": [
-      "es2016",
-      "dom"
-    ],
-    "outDir": "../dist/out-tsc-e2e",
-    "module": "commonjs",
-    "target": "es6",
-    "types": [
-      "jasmine",
-      "node"
-    ]
-  }
-}
